move deposit_radiation
